Refactor single middleware function into multiple functions that get chained together (#671)
