+++ site/src/middleware.ts
@@ -1,53 +1,21 @@
-import type { NextRequest } from "next/server";
-import { NextResponse } from "next/server";
+import { withAdminRedirectMiddleware } from "./middleware/adminRedirect";
+import { chain } from "./middleware/chain";
+import { withContentSecurityPolicyHeadersMiddleware } from "./middleware/contentSecurityPolicyHeaders";
+import { withDamRewriteMiddleware } from "./middleware/damRewrite";
+import { withDomainRewriteMiddleware } from "./middleware/domainRewrite";
+import { withPreviewMiddleware } from "./middleware/preview";
+import { withRedirectToMainHostMiddleware } from "./middleware/redirectToMainHost";
+import { withSitePreviewMiddleware } from "./middleware/sitePreview";
 
-import { configureResponse } from "./util/configureResponse";
-import { getHostByHeaders, getSiteConfigForHost, getSiteConfigs } from "./util/siteConfig";
-
-export async function middleware(request: NextRequest) {
-    const headers = request.headers;
-    const host = getHostByHeaders(headers);
-    const { pathname } = new URL(request.url);
-
-    // Preview
-    if (request.nextUrl.pathname.startsWith("/block-preview/") || request.nextUrl.pathname === "/site-preview") {
-        return NextResponse.next({ request: { headers } });
-    }
-
-    const siteConfig = await getSiteConfigForHost(host);
-    if (!siteConfig) {
-        // Redirect to Main Host
-        const redirectSiteConfig = getSiteConfigs().find(
-            (siteConfig) =>
-                siteConfig.domains.additional?.includes(host) || (siteConfig.domains.pattern && host.match(new RegExp(siteConfig.domains.pattern))),
-        );
-        if (redirectSiteConfig) {
-            return NextResponse.redirect(redirectSiteConfig.url);
-        }
-
-        throw new Error(`Cannot get siteConfig for host ${host}`);
-    }
-
-    if (pathname.startsWith("/dam/")) {
-        return NextResponse.rewrite(new URL(`${process.env.API_URL_INTERNAL}${request.nextUrl.pathname}`));
-    }
-
-    if (request.nextUrl.pathname === "/admin" && process.env.ADMIN_URL) {
-        return NextResponse.redirect(new URL(process.env.ADMIN_URL));
-    }
-
-    return configureResponse(
-        NextResponse.rewrite(
-            new URL(
-                `/${siteConfig.scope.domain}${request.nextUrl.pathname}${
-                    request.nextUrl.searchParams.toString().length > 0 ? `?${request.nextUrl.searchParams.toString()}` : ""
-                }`,
-                request.url,
-            ),
-            { request: { headers } },
-        ),
-    );
-}
+export default chain([
+    withSitePreviewMiddleware,
+    withRedirectToMainHostMiddleware,
+    withAdminRedirectMiddleware,
+    withDamRewriteMiddleware,
+    withContentSecurityPolicyHeadersMiddleware, // order matters: after redirects (that don't need csp headers), before everything else that needs csp headers
+    withPreviewMiddleware,
+    withDomainRewriteMiddleware, // must be last (rewrites all urls)
+]);
 
 export const config = {
     matcher: [
