+++ site/src/middleware/chain.ts
@@ -1,0 +1,17 @@
+import type { NextRequest } from "next/server";
+import { NextResponse } from "next/server";
+
+export type CustomMiddleware = (request: NextRequest) => NextResponse | Response | Promise<NextResponse | Response>;
+
+type MiddlewareFactory = (middleware: CustomMiddleware) => CustomMiddleware;
+// Utility function to chain multiple middlewares together
+export function chain(functions: MiddlewareFactory[], index = 0): CustomMiddleware {
+    const current = functions[index];
+
+    if (current) {
+        const next = chain(functions, index + 1);
+        return current(next);
+    }
+
+    return () => NextResponse.next();
+}
